Valida que no exista turno

+++ lib/polycon/models/appointment.rb
@@ -5,7 +5,12 @@ module Polycon
 
         def self.create(date, name)
             name=Polycon::Utils::guion(name)
-            File.open(Polycon::PATH+"#{name}"+"/"+"#{date}"+".paf", "w")
+            if(File.exist?(Polycon::PATH+"#{name}"+"/"+"#{date}"+".paf"))
+                return false
+            else
+                File.open(Polycon::PATH+"#{name}"+"/"+"#{date}"+".paf", "w")
+                return true
+            end
         end
         def initialize (name)
           self.name = name
